# perf: ThreadListCell redraw improvements