--- a/Mail/Components/ThreadCell.swift
+++ b/Mail/Components/ThreadCell.swift
@@ -80,6 +80,15 @@ struct ThreadCellDataHolder {
     }
 }
 
+extension ThreadCell: Equatable {
+    static func == (lhs: ThreadCell, rhs: ThreadCell) -> Bool {
+        return lhs.thread.id == rhs.thread.id
+            && lhs.thread.messageIds == rhs.thread.messageIds
+            && lhs.isSelected == rhs.isSelected
+            && lhs.isMultipleSelectionEnabled == rhs.isMultipleSelectionEnabled
+    }
+}
+
 struct ThreadCell: View {
     @EnvironmentObject private var mailboxManager: MailboxManager