| Module / View Type     | Count | Total Duration | Min Duration | Avg Duration | Max Duration | Std Dev Duration |
|------------------------|-------|----------------|--------------|--------------|--------------|------------------|
| Before                 |       |                |              |              |              |                  |
| ThreadListCell         | 111   | 40.87 ms       | 266.88 µs    | 368.20 µs    | 3.38 ms      | 296.85 µs        |
| ThreadListSwipeActions | 111   | 19.87 ms       | 152.62 µs    | 179.00 µs    | 510.17 µs    | 42.77 µs         |
| ThreadCell             | 19    | 18.18 ms       | 293.71 µs    | 956.76 µs    | 2.15 ms      | 778.09 µs        |
| After                  |       |                |              |              |              |                  |
| ThreadListCell         | 24    | 12.79 ms       | 305.21 µs    | 532.78 µs    | 3.36 ms      | 608.46 µs        |
| ThreadListSwipeActions | 24    | 4.86 ms        | 168.46 µs    | 202.33 µs    | 424.75 µs    | 53.99 µs         |
| ThreadCell             | 19    | 6.49 ms        | 251.12 µs    | 341.70 µs    | 471.17 µs    | 81.70 µs         |

Multiple improvements to prevent over drawing thread list cells:
- Convert some Views to POD
- Use `Equatable` if possible
- Avoid creating a realm for contacts each time

Since `Equatable` is implemented by hand for some views it may lead to bad rendering issues please test carefully.

Single View rendering isn't always faster but since we draw a lot less of them it's an overall gain.